+++ src/api/changesets/index.ts
@@ -1,6 +1,7 @@
 export * from "./createChangesetComment";
 export * from "./getChangesetDiff";
 export * from "./getChangesets";
+export * from "./subscription";
 export * from "./uploadChangeset";
 export { parseOsmChangeXml } from "./_parseOsmChangeXml";
 export { createOsmChangeXml } from "./_createOsmChangeXml";