+++ src/api/notes/index.ts
@@ -1,2 +1,3 @@
 export * from "./getNotes";
 export * from "./noteActions";
+export * from "./subscription";