+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 2.4.0 (2025-01-16)
+
 - Added a method to easily switch users (logout & log back in)
 
 ## 2.3.0 (2024-12-03)